Add anEmbeddedPostgresqlTest junit test
This tests spinning up an embedded PostgreSQL cluster instance,
connecting to it and doing a basic select test.  It then uses Flyway to
run all of the DDL SQL scripts to create the database schema, and reads
from the tables, and tests inserting to the customers table.  More to come.