Use correct fields in the CustomerQuotaReducer
To get quotas to reduce correctly, use prefixed Customer fields and
plain Quota fields.  In accumulate(), build a list of quotas.  The first
quota to be added is added in the CustomerMapper, but subsequent quotas
for the same customer id are appended to the quotas list of the existing
Customer instance.