Add the Quota.subject field to enable quotas without customer ids
Quotas can be unassociated with customers or subjects (like an ORCID
or a group DN), or associated with one or the other.  We need to be
able to set quootas based on subject alone (in the case of a group DN),
where the group DN isn't bound to a particular customer.
Change the Quota model to add the subject field, and update the QuotaTest
to reflect the change.  Update the SQL migrations to add the column in
the database, and update the jdbi store classes to reflect the change.
Update store tests, mappers, and helpers to reflect the quota change.
Add a QuotaHelper.insertTestQuotaWithSubject() method to help with
testing quotas with just subjects defined.