Reorganize SQL DDL scripts to add subscriptions
Add a subscriptions table and create a foreign key constraint from
the quotas table to associate quotas with a subscription.
Also create a foreign key constraint on the customers
table id field to associate a subscription with a customer.
Reorganize the scripts to add the subscriptions table in the correct order.