Add a SubscriptionMapper class to map subscription rows to instances.
Given a resultset row from the SubscriptionStore interface, map
columns into a Subscription instance.  If there are product columns,
create a Product instance.  If there are Quota columns, create a Quota
instance, and add it to a quotas list.  The calling row reducer should
keep adding to the quotas list.