Type softLimit, hardLimit, and usage fields as double precision:
I realized that when I moved to Integer and required settings in units like megabytes
for storage quotas (since Jackson was duck typing to IntNode), we would be unable to
store usage values below 1 megabyte.  For files of say, 500kb, we're out of luck.

So know we require all quota and usage numbers to be floating point, e.g. serialized
as 0.5 (for 500kb in megabytes), or 1.0 (for a portal count quota).