Improve calls to listOrders():
Handle admin calls separately from non-admin calls, and always list orders for admins.
When no orders exist, return 404 NotFound.
For non-admins, return orders if the caller subject equals the customer subject - i.e. they own the order.
When passing in a list of subjects, redact orders unless the customer subject matches the caller subject. We want to expose financial portions of the order to owners, but send back other order details pertinent to order group members. Note that we need to not redact when the caller subject is a group owner. This is a future TODO.

refs #70